Add MLX5DV_SIG_CHECK_COPY_EN_COMP_MASK comp_mask support
Support the "MLX5DV_SIG_CHECK_COPY_EN_COMP_MASK" comp_mask of a signature
attribute, which is the preferred way to configure a block signature.

This flag can be enabled with "-m" or "--enable-check-copy-en-mask"; When
it is enabled, the check_copy_en field specifies how the guard and tags
are checked in input domain, and how they are copied from the source domain
to the destination domain, and the check_mask and copy_mask fields are
ignored.

This flag is supported only when we have the MLX5DV_SIG_PROT_CAP_NVMEDIF
capability.

Signed-off-by: Mark Zhang <markzhang@nvidia.com>